refactor bindings into separate file
